Store new emojis directly in collection
Instead of storing a new emoji to a variable, and then creating a
dictionary that maps to that variable, this stores the new emoji
directly in the dictionary. This simplifies development by requiring the
addition of only one line to add an emoji instead of two. Contributing
guidelines adjusted for this change.